adding a basic layout of editing deck
