+++ src/App.tsx
@@ -10,8 +10,7 @@ import OauthRedirect from "./OAuthRedirect/OauthRedirect";
 import {SetChoice} from "./SetChoice/SetChoice";
 import PrivateRoute from "./PrivateRoute/PrivateRoute";
 import WaitingRoom from "./WaitingRoom/WaitingRoom";
-import {EditDeckQuestions} from './EditDeck/EditDeckQuestions';
-import {LoadAndEditDeck} from "./EditDeck/LoadAndEditDeck";
+import {EditDeck} from "./EditDeck";
 
 function App() {
     document.title = "Flashcard Bonanza";
@@ -26,7 +25,7 @@ function App() {
                 <Route path="/in-game" element={<PrivateRoute><InGame/></PrivateRoute>}/>
                 <Route path="/summary" element={<PrivateRoute><ScoreSummary/></PrivateRoute>}/>
                 <Route path="/waiting-room" element={<PrivateRoute><WaitingRoom/></PrivateRoute>}/>
-                <Route path="/edit-deck" element={<PrivateRoute><LoadAndEditDeck/></PrivateRoute>}/>
+                <Route path="/edit-deck" element={<PrivateRoute><EditDeck/></PrivateRoute>}/>
             </Routes>
         </div>
     );
