+++ src/EditDeck/index.tsx
@@ -1,0 +1,1 @@
+export {LoadAndEditDeck as EditDeck} from "./LoadAndEditDeck";
