+++ src/EditDeck/old/DeckElement.tsx
@@ -2,7 +2,7 @@ import AccordionItem from "react-bootstrap/AccordionItem";
 import {QuestionList} from "./QuestionList";
 import {Accordion} from "react-bootstrap";
 import React from "react";
-import {Deck} from "../types/BackendModels";
+import {Deck} from "../../types/BackendModels";
 
 export const DeckElement: React.FC<{ deck: Deck, idx: number }> = ({deck, idx}) => (
     <Accordion key={idx}>
