+++ src/EditDeck/old/SelectedDeck.tsx
@@ -1,5 +1,5 @@
 import React, {useEffect, useState} from "react";
-import {Deck, FlashCard} from "../types/BackendModels";
+import {Deck, FlashCard} from "../../types/BackendModels";
 import {Button, Col, Form, FormControl, Row} from "react-bootstrap";
 // import {IdHeader} from "./formElements/IdHeader";
 // import {FormDescription} from "./formElements/FormDescription";
